♻️ refactor: dissolve the private half of galax.utils
`galax.utils` is a shared grab-bag every subpackage reaches into. Each private
module in it has a single consuming subpackage, so each moves there and the
cross-subpackage coupling goes away.

  - `_jax.vectorize_method` -> `galax.potential._src.jax` (sole consumer)
  - `_boundinstance`        -> `galax._boundinstance`
  - `_shape`                -> split, see below

`_shape` served three subpackages, so it splits by function rather than moving
whole: `batched_shape` / `vector_batched_shape` to
`galax.coordinates._src.shape`, the lowest subpackage using them, and
`expand_arr_dims` / `expand_batch_dims` to `galax.potential._src.shape`, their
only consumer. `_jax.quaxify` goes with `batched_shape`, since coordinates is
the lowest subpackage that decorates with it.

`quaxify` looks like a pointless three-line wrapper over `quax.quaxify` -- this
codebase already imports `quax.quaxify` directly in
`coordinates/_src/{psps,pscs}/register_vectorapi.py` -- but its `cast` is
load-bearing. `quax.quaxify` is untyped, so decorating with it directly makes
the decorated function untyped; dropping the wrapper made `batched_shape`,
`expand_batch_dims` and `expand_arr_dims` untyped and added three mypy errors.

The public half of `galax.utils` is untouched: `dataclasses` and
`loop_strategies` stay put. Moving those changes public API, so it belongs with
the breaking release rather than ahead of it.

Tests follow their subjects: `tests/unit/utils/test_shape.py` splits across
coordinates and potential, `test_jax.py` moves to potential.

Split out of #799.
