+++ src/galax/coordinates/_src/shape.py
@@ -1,9 +1,14 @@
-"""galax: Galactic Dynamix in Jax."""
+"""Shape utilities. Private API.
+
+`coordinates` is the lowest portion that uses these, so they live here and
+`potential` / `dynamics` import them from here.
+"""
 
 __all__: tuple[str, ...] = ()
 
+from collections.abc import Callable
 from jaxtyping import Array, Shaped
-from typing import Literal, TypeAlias, overload
+from typing import Any, Literal, TypeAlias, cast, overload
 
 import quax
 
@@ -11,10 +16,19 @@ import coordinax as cx
 import quaxed.numpy as jnp
 
 import galax._custom_types as gt
-from galax.utils._jax import quaxify
 
 AnyScalar: TypeAlias = Shaped[Array, ""]
 ArrayAnyShape: TypeAlias = Shaped[Array | quax.ArrayValue, "..."]
+
+
+def quaxify[**P, R](fn: Callable[P, R], *, filter_spec: Any = True) -> Callable[P, R]:
+    """Wrap a function with `quax.quaxify`, preserving its signature.
+
+    `quax.quaxify` is untyped, so decorating with it directly makes the
+    decorated function untyped too. The `cast` keeps type checkers seeing the
+    original signature.
+    """
+    return cast("Callable[P, R]", quax.quaxify(fn, filter_spec=filter_spec))
 
 
 def vector_batched_shape(obj: cx.vecs.AbstractVector, /) -> tuple[gt.Shape, int]:
@@ -71,7 +85,7 @@ def batched_shape(
     Standard imports:
 
         >>> import quaxed.numpy as jnp
-        >>> from galax.utils._shape import batched_shape
+        >>> from galax.coordinates._src.shape import batched_shape
 
     Expecting a scalar:
 
@@ -105,76 +119,3 @@ def batched_shape(
     shape: gt.Shape = jnp.asarray(arr).shape
     ndim = len(shape)
     return shape[: ndim - expect_ndim], shape[ndim - expect_ndim :]
-
-
-@quaxify
-def expand_batch_dims(arr: ArrayAnyShape, /, ndim: int) -> ArrayAnyShape:
-    """Expand the batch dimensions of an array.
-
-    Parameters
-    ----------
-    arr : array-like
-        The array to expand the batch dimensions of.
-    ndim : int
-        The number of batch dimensions to expand.
-
-    Returns
-    -------
-    arr : array-like
-        The array with expanded batch dimensions.
-
-    Examples
-    --------
-    >>> import jax.numpy as jnp
-    >>> from galax.utils._shape import expand_batch_dims
-
-    >>> expand_batch_dims(jnp.array(0), ndim=0).shape
-    ()
-
-    >>> expand_batch_dims(jnp.array([0]), ndim=0).shape
-    (1,)
-
-    >>> expand_batch_dims(jnp.array(0), ndim=1).shape
-    (1,)
-
-    >>> expand_batch_dims(jnp.array([0, 1]), ndim=1).shape
-    (1, 2)
-    """
-    return jnp.expand_dims(arr, axis=tuple(range(ndim)))
-
-
-@quaxify
-def expand_arr_dims(arr: ArrayAnyShape, /, ndim: int) -> ArrayAnyShape:
-    """Expand the array dimensions of an array.
-
-    Parameters
-    ----------
-    arr : array-like
-        The array to expand the array dimensions of.
-    ndim : int
-        The number of array dimensions to expand.
-
-    Returns
-    -------
-    arr : array-like
-        The array with expanded array dimensions.
-
-    Examples
-    --------
-    >>> import jax.numpy as jnp
-    >>> from galax.utils._shape import expand_arr_dims
-
-    >>> expand_arr_dims(jnp.array(0), ndim=0).shape
-    ()
-
-    >>> expand_arr_dims(jnp.array([0]), ndim=0).shape
-    (1,)
-
-    >>> expand_arr_dims(jnp.array(0), ndim=1).shape
-    (1,)
-
-    >>> expand_arr_dims(jnp.array([0, 0]), ndim=1).shape
-    (2, 1)
-    """
-    nbatch = len(arr.shape)
-    return jnp.expand_dims(arr, axis=tuple(nbatch + i for i in range(ndim)))
