+++ src/galax/dynamics/_src/legacy/interp_psp.py
@@ -19,8 +19,8 @@ import galax._custom_types as gt
 import galax.coordinates as gc
 from galax.coordinates._src.frames import SimulationFrame
 from galax.coordinates._src.pscs.base import ComponentShapeTuple
+from galax.coordinates._src.shape import batched_shape, vector_batched_shape
 from galax.dynamics._src.orbit import PhaseSpaceInterpolation
-from galax.utils._shape import batched_shape, vector_batched_shape
 
 
 @final
