+++ src/galax/dynamics/_src/orbit/plot_helper.py
@@ -13,7 +13,7 @@ from plum import PromisedType, dispatch
 
 from plotting_backends import AbstractPlottingBackend, MatplotlibBackend
 
-from galax.utils._boundinstance import BndTo, InstanceDescriptor
+from galax._boundinstance import BndTo, InstanceDescriptor
 
 ProxyAbstractOrbit = PromisedType("AbstractOrbit")
