+++ src/galax/potential/_src/builtin/monari2016.py
@@ -20,9 +20,9 @@ from unxt.quantity import AllowValue
 
 import galax._custom_types as gt
 from galax.potential._src.base_single import AbstractSinglePotential
+from galax.potential._src.jax import vectorize_method
 from galax.potential._src.params.base import AbstractParameter
 from galax.potential._src.params.field import ParameterField
-from galax.utils._jax import vectorize_method
 
 
 @final
