+++ src/galax/potential/_src/plot.py
@@ -13,7 +13,7 @@ from plum import PromisedType, dispatch
 
 from plotting_backends import AbstractPlottingBackend, MatplotlibBackend
 
-from galax.utils._boundinstance import BndTo, InstanceDescriptor
+from galax._boundinstance import BndTo, InstanceDescriptor
 
 ProxyAbstractPotential = PromisedType("AbstractPotential")
