+++ tests/unit/potential/test_shape.py
@@ -1,4 +1,4 @@
-"""Test the `galax.utils._shape` module."""
+"""Test the `galax.potential._src.shape` module."""
 
 import jax
 import pytest
@@ -6,41 +6,11 @@ import pytest
 import quaxed.numpy as jnp
 import unxt as u
 
-import galax._custom_types as gt
-from galax.utils._shape import batched_shape, expand_arr_dims, expand_batch_dims
-
-
-class TestBatchedShape:
-    """Test the `galax.utils._shape.batched_shape` function."""
-
-    @pytest.mark.parametrize(
-        ("arr", "expect_ndim", "expect"),
-        [
-            # ArrayLike
-            (jnp.asarray(42), 0, ((), ())),
-            (jnp.asarray([1]), 0, ((1,), ())),
-            (jnp.asarray([1, 2, 3]), 1, ((), (3,))),
-            (jnp.asarray([[1, 2], [3, 4]]), 1, ((2,), (2,))),
-            (jnp.asarray([[1, 2], [3, 4]]), 2, ((), (2, 2))),
-            # Quantity
-            (u.Q(42, "m"), 0, ((), ())),
-            (u.Q([1], "m"), 0, ((1,), ())),
-            (u.Q([1, 2, 3], "m"), 1, ((), (3,))),
-            (u.Q([[1, 2], [3, 4]], "m"), 1, ((2,), (2,))),
-            (u.Q([[1, 2], [3, 4]], "m"), 2, ((), (2, 2))),
-        ],
-    )
-    def test_batched_shape(
-        self, arr: jax.Array, expect_ndim: int, expect: tuple[gt.Shape, gt.Shape]
-    ) -> None:
-        """Test the `galax.utils._shape.batched_shape` function."""
-        batch, shape = batched_shape(arr, expect_ndim=expect_ndim)
-        assert batch == expect[0]
-        assert shape == expect[1]
+from galax.potential._src.shape import expand_arr_dims, expand_batch_dims
 
 
 class TestExpandBatchDims:
-    """Test :func:`galax.utils._shape.expand_batch_dims`."""
+    """Test :func:`galax.potential._src.shape.expand_batch_dims`."""
 
     @pytest.mark.parametrize(
         ("arr", "ndim", "expect"),
@@ -66,14 +36,14 @@ class TestExpandBatchDims:
     def test_expand_batch_dims(
         self, arr: jax.Array, ndim: int, expect: jax.Array
     ) -> None:
-        """Test :func:`galax.utils._shape.expand_batch_dims`."""
+        """Test :func:`galax.potential._src.shape.expand_batch_dims`."""
         got = expand_batch_dims(arr, ndim=ndim)
         assert jnp.array_equal(got, expect)
         assert got.shape == expect.shape
 
 
 class TestExpandArrDims:
-    """Test :func:`galax.utils._shape.expand_arr_dims`."""
+    """Test :func:`galax.potential._src.shape.expand_arr_dims`."""
 
     @pytest.mark.parametrize(
         ("arr", "ndim", "expect"),
@@ -99,7 +69,7 @@ class TestExpandArrDims:
     def test_expand_arr_dims(
         self, arr: jax.Array, ndim: int, expect: jax.Array
     ) -> None:
-        """Test :func:`galax.utils._shape.expand_arr_dims`."""
+        """Test :func:`galax.potential._src.shape.expand_arr_dims`."""
         got = expand_arr_dims(arr, ndim=ndim)
         assert jnp.array_equal(got, expect)
         assert got.shape == expect.shape
